+++ wxcadm/device.py
@@ -320,3 +320,223 @@ class DeviceMember:
         self.remote_ip: Optional[str] = member_info.get('remoteIP', None)
         self.line_port: Optional[str] = member_info.get('linePort', None)
 
+
+class DeviceList(UserList):
+    _endpoint = "v1/devices"
+    _endpoint_items_key = None
+    _item_endpoint = "v1/devices/{item_id}"
+    _item_class = Device
+
+    def __init__(self, parent: Union[wxcadm.Org, wxcadm.Location]):
+        super().__init__()
+        log.debug("Initializing DeviceList")
+        self.parent: wxcadm.Org | wxcadm.person.Person | wxcadm.workspace.Workspace | wxcadm.Location = parent
+        self.data: list = self._get_data()
+
+    def _get_data(self) -> list:
+        log.debug("_get_data() started")
+        params = {}
+
+        if isinstance(self.parent, wxcadm.Org):
+            log.debug(f"Using Org ID {self.parent.id} as data filter")
+            params['orgId'] = self.parent.id
+        elif isinstance(self.parent, wxcadm.Location):
+            log.debug(f"Using Location ID {self.parent.id} as data filter")
+            params['locationId'] = self.parent.id
+        elif isinstance(self.parent, wxcadm.person.Person):
+            log.debug(f"Using API endpoint v1/telephony/config/people/{self.parent.id}/devices")
+            self._endpoint = f"v1/telephony/config/people/{self.parent.id}/devices"
+            self._endpoint_items_key = 'devices'
+        elif isinstance(self.parent, wxcadm.workspace.Workspace):
+            log.debug(f"Using Workspace ID {self.parent.id} as data filter")
+            params['workspaceId'] = self.parent.id
+        else:
+            log.warn("Parent class is not Org or Location, so all items will be returned")
+        response = webex_api_call('get', self._endpoint, params=params)
+        items = []
+        if self._endpoint_items_key is not None:
+            log.info(f"Found {len(response[self._endpoint_items_key])} items")
+            for entry in response[self._endpoint_items_key]:
+                items.append(self._item_class(parent=self.parent, config=entry))
+        else:
+            log.info(f"Found {len(response)} items")
+            for entry in response:
+                items.append(self._item_class(parent=self.parent, config=entry))
+        return items
+
+    def refresh(self):
+        """ Refresh the list of instances from Webex
+
+        Returns:
+            bool: True on success, False otherwise.
+
+        """
+        self.data = self._get_data()
+        return True
+
+    def get(self, id: Optional[str] = None, name: Optional[str] = None, spark_id: Optional[str] = None):
+        """ Get the instance associated with a given ID, Name, or Spark ID
+
+        Only one parameter should be supplied in normal cases. If multiple arguments are provided, the Locations will be
+        searched in order by ID, Name, and finally Spark ID. If no arguments are provided, the method will raise an
+        Exception.
+
+        Args:
+            id (str, optional): The Call Queue ID to find
+            name (str, optional): The Call Queue Name to find. Case-insensitive.
+            spark_id (str, optional): The Spark ID to find
+
+        Returns:
+            CallQueue: The CallQueue instance correlating to the given search argument.
+
+        Raises:
+            ValueError: Raised when the method is called with no arguments
+
+        """
+        if id is None and name is None and spark_id is None:
+            raise ValueError("A search argument must be provided")
+        if id is not None:
+            for item in self.data:
+                if item.id == id:
+                    return item
+        if name is not None:
+            for item in self.data:
+                if item.display_name.lower() == name.lower():
+                    return item
+        if spark_id is not None:
+            for item in self.data:
+                if item.spark_id == spark_id:
+                    return item
+        return None
+
+    def create(self, model: str,
+               mac: Optional[str] = None,
+               password: Optional[str] = None,
+               person: Optional[wxcadm.person.Person] = None,
+               workspace: Optional[wxcadm.workspace.Workspace] = None):
+        """ Add a new device to the Workspace, Person or Org
+
+        In order to use this method, you must know the model of the device that you are adding, as expected by the
+        Webex API. If you are adding a "Generic IPPhone Customer Managed" device, you can use that value or simply
+        send ``model='GENERIC'`` as an alias. You can find the full list of models with the
+        :py:meth:`Org.get_supported_devices()` method.
+
+        If the MAC address is passed, a device will be created with the provided MAC address. If no MAC address is
+        passed, an Activation Code will be generated and returned as part of the response. Your integration/token must
+        have the ``identity:placeonetimepassword_create`` scope to create Activation Codes for devices.
+
+        Args:
+            model (str): The model name of the device being added
+
+            mac (str, optional): The MAC address of the device being added
+
+            password (str, optional): Only valid when creating a Generic IPPhone Customer Managed device. If a
+                password is not provided, the Webex API will generate a unique, compliant SIP password and return it
+                in the response.
+
+            person (Person, optional): If the :class:`DeviceList` was accessed via :attr:`Org.devices`, a Person or
+                Workspace must be provided when creating a Device
+
+            workspace (Workspace, optional): If the :class:`DeviceList` was accessed via :attr:`Org.devices`, a
+                Workspace or Person must be provided when creating a Device
+
+        Returns:
+            dict: The dict values vary based on the type of device being activated. If the device fails for any reason,
+                False will be returned. At the moment, Webex doesn't provide very useful failure reasons, but those may
+                be added to the return value when they are available.
+
+        Raises:
+            ValueError: Raised when the DeviceList cannot determine which Person or Workspace to add the device to.
+                This is normally when the DeviceList was created at the Org level with :attr:`Org.devices`. Ensure you
+                are passing a ``workspace`` or ``person`` argument to the method.
+
+        """
+        log.info(f"Adding a device to {self.parent.name}")
+        if isinstance(self.parent, wxcadm.workspace.Workspace):
+            payload = {
+                "workspaceId": self.parent.id,
+                "model": model
+            }
+        elif isinstance(self.parent, (wxcadm.person.Person, wxcadm.person.Me)):
+            payload = {
+                "personId": self.parent.id,
+                "model": model
+            }
+        elif isinstance(self.parent, wxcadm.org.Org):
+            if person is not None:
+                payload = {
+                    "personId": person.id,
+                    "model": model
+                }
+            elif workspace is not None:
+                payload = {
+                    "workspaceId": workspace.id,
+                    "model": model
+                }
+            else:
+                raise ValueError("Device or Workspace must be provided")
+
+        data_needed = False  # Flag that we need to get platform data once we have a Device ID
+        if mac is None and model != 'Imagicle Customer Managed':
+            # If no MAC address is provided, just generate an activation code for the device
+            try:
+                response = webex_api_call('post',
+                                          'v1/devices/activationCode',
+                                          payload=payload)
+                log.debug(f"\t{response}")
+            except APIError:
+                return False
+
+            # Get the ID of the device we just inserted
+            device_id = response.get('id', None).replace('=', '')
+
+            results = {
+                'device_id': device_id,
+                'activation_code': response['code']
+            }
+        else:
+            payload['mac'] = mac
+            if model.upper() == "GENERIC" or model == "Generic IPPhone Customer Managed" or model == 'Imagicle ' \
+                                                                                                     'Customer Managed':
+                if payload['model'] != 'Imagicle Customer Managed':
+                    payload[
+                        'model'] = "Generic IPPhone Customer Managed"  # Hard-code what the API expects (for now)
+                data_needed = True
+                if password is None:  # Generate a unique password
+                    if isinstance(self.parent, (wxcadm.person.Person, wxcadm.person.Me)):
+                        password_location = self.parent.location
+                    if isinstance(self.parent, wxcadm.workspace.Workspace):
+                        password_location = self.parent._parent.locations.webex_calling(single=True).id
+                    if isinstance(self.parent, wxcadm.org.Org):
+                        password_location = self.parent.locations.webex_calling(single=True).id
+                    response = webex_api_call('POST',
+                                              f'v1/telephony/config/locations/{password_location}/actions/'
+                                              f'generatePassword/invoke')
+                    password = response['exampleSipPassword']
+                payload['password'] = password
+            response = webex_api_call('post', 'v1/devices', payload=payload)
+            log.debug(f"\t{response}")
+
+            # Get the ID of the device we just inserted
+            device_id = response.get('id', None).replace('=', '')
+
+            results = {
+                'device_id': device_id,
+                'mac': response['mac']
+            }
+
+            if data_needed is True:
+                response = webex_api_call('get', f'/v1/telephony/config/devices/{device_id}')
+                results['sip_auth_user'] = response['owner']['sipUserName']
+                results['line_port'] = response['owner']['linePort']
+                results['password'] = password
+                results['sip_userpart'] = response['owner']['linePort'].split('@')[0]
+                results['sip_hostpart'] = response['owner']['linePort'].split('@')[1]
+                results['sip_outbound_proxy'] = response['proxy']['outboundProxy']
+                results['sip_outbound_proxy_srv'] = f"_sips._tcp.{response['proxy']['outboundProxy']}"
+
+        # Provide the Device instance in the response as well
+        self.refresh()
+        results['device_object'] = self.get(id=device_id)
+        return results
+